remove ast-types as dependency. Access via recast.types.
recast adds a `File` type to `ast-types`.

Accessing `ast-types` by setting it as a dependency and then doing:
`require("ast-types")` used to work, but started throwing
`AssertionError: did not recognize object of type "file"`.

Not sure why this is happening.
1. Type defs used to be global, now they are not?
2. Semver incompatible versions so npm was loading two separate copies?
3. Something else?

Either way - accessing via: `require("recast").types` solves the issue.
This is probably the correct way to do this and I just didn't know better.