example(proxyDirective): ensure it works without transclude
angular normally throws if you request an injection that does not
exist. $transclude exists conditionally (depending on whether
the directive in question uses transclusion). @proxyDirectiveController
always requests the $transclude injection regardless of the underlying
directive, so it could potentially cause angular to throw. Furtuneately,
those angular guys are smart and $transclude is an exception to the
"throws if undefined rule". This test proves that.